Update README.md
The market research team at AeroFit wants to identify the characteristics of the target audience
for each type of treadmill offered by the company, to provide a better recommendation of the
treadmills to the new customers. The team decides to investigate whether there are differences
across the product with respect to customer characteristics.
1. Perform descriptive analytics to create a customer profile for each AeroFit treadmill product
by developing appropriate tables and charts.
2. For each AeroFit treadmill product, construct two-way contingency tables and compute all
conditional and marginal probabilities along with their insights/impact on the business.